Move packets and common.go to internal module
