+++ src/internal/common.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import "time"
 
@@ -10,3 +10,8 @@ const PORT = "8888"
 const KAFKA_ADDRESS = "localhost" // Use Docker service name
 const KAFKA_PORT = "9092"
 const MESSAGE_TIMEOUT = 5 * time.Second
+
+// Redis constants
+const REDIS_ADDRESS = "localhost" // Use Docker service name
+const REDIS_PORT = "6379"
+const REDIS_DYNAMIC_DATA_TTL = 10
